Finished importing the report. Requires cleanup and iterating.
