+++ KenticoInspector.Reports/WebPartSecurityAnalysis/GetPageTemplateWebParts.sql
