+++ KenticoInspector.Reports/WebPartSecurityAnalysis/PageTemplate.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace KenticoInspector.Reports.WebPartSecurityAnalysis
+{
+    public class PageTemplate
+    {
+        public string PageTemplateDisplayName { get; set; }
+        public string PageTemplateWebParts { get; set; }
+    }
+}
